// src/app/models/signup-post.interface.ts
export interface SignupPost {
  UserID?: string;
  Nationality?: string;
  Title?: string;
  Name?: string;
  CivilID?: string;
  ValidDt?: string;
  Email?: string;
  MobileNo?: string;
  S_Question_1?: string;
  S_Answer_1?: string;
  S_Question_2?: string;
  S_Answer_2?: string;
  S_Question_3?: string;
  S_Answer_3?: string;
  S_Question_4?: string;
  S_Answer_4?: string;
  S_Question_5?: string;
  S_Answer_5?: string;
  AvatarID?: string;
  PhraseID?: string;
  Password?: string;
  OTP?: string;
}